pyproject.toml
[project]
name = "agento"
version = "0.1.0"
description = "CLI/TUI AI agent for code generation and DevOps automation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Agento Team", email = "hello@agento.dev"}
]
requires-python = ">=3.11"
keywords = ["ai", "agent", "cli", "tui", "code-generation", "devops"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core LangGraph
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",

    # LLM Providers
    "openai>=1.12.0",
    "httpx>=0.27.0",

    # UI
    "rich>=13.7.0",
    "textual>=0.80.0",

    # Data
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",

    # Linting & Formatting
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.9.0",

    # Type stubs
    "types-requests>=2.31.0",
]

memory = [
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]

mcp = [
    "mcp>=1.0.0",
]

docker = [
    "docker>=7.0.0",
]

all = [
    "agento[dev,memory,mcp,docker]",
]

colab = [
    "google-colab>=1.0.0",
]

[project.scripts]
agent = "agento.main:run"
agento = "agento.main:run"

[project.urls]
Homepage = "https://github.com/ajdevy/agento"
Documentation = "https://github.com/ajdevy/agento#readme"
Repository = "https://github.com/ajdevy/agento"
Issues = "https://github.com/ajdevy/agento/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
ignore_errors = false

[[tool.mypy.overrides]]
module = "langgraph.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "langchain.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "agento.core.nodes"
ignore_errors = true

[[tool.mypy.overrides]]
module = "agento.application.agent"
ignore_errors = true

[[tool.mypy.overrides]]
module = "agento.application.pipeline"
ignore_errors = true

[[tool.mypy.overrides]]
module = "agento.infrastructure.memory.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "agento.domain.services.memory_service"
ignore_errors = true

[tool.coverage.run]
source = ["src/agento"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.fail-under]
min = 95

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=src/agento --cov-report=term-missing --cov-fail-under=95"
